backend/requirements_updated.txt
langchain>=0.1.7
langchain-community
langchain-core>=0.1.7,<0.2
langchain-groq
langchain-google-genai
langchain-chroma
langchain-openai
langchain>=0.1.7
langchain-community
langchain-core>=0.1.7,<0.2
langchain-groq
langchain-google-genai
langchain-chroma
langchain-openai
pydantic
pytest
black
flake8
python-dotenv
fitz
pillow
pytesseract